feat(longhorn): generalize install for non-AWS providers
Lift the Longhorn install logic out of the AWS provider into a shared
package so the existing-cluster and hetzner providers can opt in. On
clusters without managed RWX (Hetzner hcloud-volumes is RWO-only,
on-prem k3s, kind/k3d), charts that need RWX previously had to fall
back to the in-cluster NFS-on-RWO workaround, which is fragile on
overlayfs hosts. Closes #269.

Shared package pkg/storage/longhorn:
- Install(ctx, kubeconfigBytes, *Config) — idempotent helm install or
  upgrade. iSCSI prerequisite DaemonSet runs on first install only;
  upgrade path skips it (saves the 3-minute readiness wait).
- Config + IsEnabled / Replicas helpers. nil *Config = do not install.

aws provider:
- Longhorn field type is now *longhorn.Config; aws.LonghornConfig is a
  type alias so existing yaml under cluster.aws.longhorn keeps working.
- LonghornEnabled() retains AWS's nil-block-as-enabled default.
- Provider.Deploy calls longhorn.Install directly.

existing provider:
- New opt-in longhorn block (nil = no install).
- Deploy installs Longhorn when enabled.
- GetStorageClass / InfraSettings.StorageClass returns 'longhorn' when
  Longhorn is enabled and storage_class is unset; explicit
  storage_class always wins.

hetzner provider:
- Same opt-in. Deploy reads the kubeconfig hetzner-k3s wrote and
  installs Longhorn after cluster create.
- InfraSettings.StorageClass flips from hcloud-volumes to longhorn
  when enabled.
