+++ pkg/provider/aws/config.go
@@ -1,5 +1,22 @@
 package aws
 
+import "github.com/nebari-dev/nebari-infrastructure-core/pkg/storage/longhorn"
+
+// LonghornConfig is kept as a package-local alias so existing yaml under
+// `longhorn:` still unmarshals into the same shape; the underlying type now
+// lives in pkg/storage/longhorn so non-AWS providers can share install logic.
+type LonghornConfig = longhorn.Config
+
+// Helpers for tests that need to construct a LonghornConfig without exporting
+// internals from the shared package.
+func newLonghornCfgWithEnabled(enabled *bool) *LonghornConfig {
+	return &LonghornConfig{Enabled: enabled}
+}
+
+func newLonghornCfgWithReplicas(n int) *LonghornConfig {
+	return &LonghornConfig{ReplicaCount: n}
+}
+
 type Config struct {
 	Region                   string               `yaml:"region"`
 	StateBucket              string               `yaml:"state_bucket,omitempty"`
@@ -19,7 +36,7 @@ type Config struct {
 	NodeGroups               map[string]NodeGroup `yaml:"node_groups"`
 	Tags                     map[string]string    `yaml:"tags,omitempty"`
 	EFS                      *EFSConfig           `yaml:"efs,omitempty"`
-	Longhorn                 *LonghornConfig      `yaml:"longhorn,omitempty"`
+	Longhorn                 *longhorn.Config     `yaml:"longhorn,omitempty"`
 }
 
 type NodeGroup struct {
@@ -41,25 +58,20 @@ type Taint struct {
 }
 
 // LonghornEnabled returns whether Longhorn distributed block storage should
-// be deployed on this AWS cluster. Defaults to true when the Longhorn config
-// is nil or Enabled is not set.
+// be deployed on this AWS cluster. Defaults to true when the Longhorn block
+// is omitted entirely — Longhorn is the AWS storage default. The shared
+// longhorn.Config defaults to disabled-when-nil because non-AWS providers
+// require an explicit opt-in.
 func (c *Config) LonghornEnabled() bool {
 	if c.Longhorn == nil {
 		return true
 	}
-	if c.Longhorn.Enabled == nil {
-		return true
-	}
-	return *c.Longhorn.Enabled
+	return c.Longhorn.IsEnabled()
 }
 
 // LonghornReplicaCount returns the number of Longhorn volume replicas.
-// Defaults to 2 when not set.
 func (c *Config) LonghornReplicaCount() int {
-	if c.Longhorn == nil || c.Longhorn.ReplicaCount == 0 {
-		return 2
-	}
-	return c.Longhorn.ReplicaCount
+	return c.Longhorn.Replicas()
 }
 
 type EFSConfig struct {
@@ -82,11 +94,4 @@ func (c *Config) EFSStorageClassName() string {
 		return defaultEFSStorageClassName
 	}
 	return c.EFS.StorageClassName
-}
-
-type LonghornConfig struct {
-	Enabled        *bool             `yaml:"enabled,omitempty"`
-	ReplicaCount   int               `yaml:"replica_count,omitempty"`
-	DedicatedNodes bool              `yaml:"dedicated_nodes,omitempty"`
-	NodeSelector   map[string]string `yaml:"node_selector,omitempty"`
 }
